requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
psycopg[binary]==3.2.3
python-dotenv==1.0.0
gunicorn==21.2.0
